requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pillow>=10.0.0
numpy>=1.26.0
opencv-python>=4.10.0.84
anomalib>=2.1.0
python-dotenv>=1.0.1
